refactor: moved hashPassword
- Moved hashPassword from Login container to 'hash.js' file
