+++ UI/app/containers/Login/index.js
@@ -39,12 +39,15 @@ import {
 import {
   Link,
   useLocation,
-} from "react-router-dom";
+} from 'react-router-dom';
 import { useCookies } from 'react-cookie';
 import { useDispatch, useSelector } from "react-redux";
 
 import { makeStyles, useTheme } from '@material-ui/core/styles';
 
+import {
+  hashPassword
+} from '../../utils/hash'
 
 const drawerWidth = 0;
 
@@ -74,29 +77,6 @@ const useStyles = makeStyles((theme) => ({
  },
 }));
 
-// import hash function argon2id for password
-const {argon2id} = window.hashwasm;
-// Salt should be at least 8 bytes long
-const salt = 'Cerberosï\x8Dzß~9ão'
-const hashPassword = async (username, password) => {
-    // salt + username bit level xor
-    const _salt = Array.from(salt)
-    .map((item, idx) => username[idx] ?
-      String.fromCharCode(item.charCodeAt(0) ^ username[idx].charCodeAt(0))
-      : item)
-    .join('')
-  const key = await argon2id({
-    password,
-    salt: _salt, // salt is a buffer containing random bytes
-    parallelism: 1,
-    iterations: 128,
-    memorySize: 256, // use 512KB memory
-    hashLength: 32, // output size = 32 bytes
-    outputType: 'encoded', // return standard encoded string containing parameters needed to verify the key
-  });
-  return key
-}
-
 export function Login({
   token,
   loginState,
@@ -106,7 +86,6 @@ export function Login({
 }) {
 
   // let location = useLocation()
-  // debugger
   const classes = useStyles();
   const theme = useTheme();
